+++ tests/performance/test_performance.c
@@ -25,7 +25,7 @@ static double get_time_ms(void) {
 static double get_time_ms(void) {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return tv.tv_sec * 1000.0 + tv.tv_usec / 1000.0;
+    return (double)tv.tv_sec * 1000.0 + (double)tv.tv_usec / 1000.0;
 }
 #endif
